refactor: extract Manager abstraction from types.ts to abstractions/Manager.ts
Moves the Manager DI abstraction out of the shared types file into its
own file under abstractions/, consistent with other abstractions in the package.
