+++ packages/api-elasticsearch-tasks/src/tasks/dataSynchronization/elasticsearch/ElasticsearchSynchronize.ts
@@ -71,7 +71,7 @@ class ElasticsearchSynchronizeImpl implements Abstraction.Interface {
         });
 
         const elasticsearchSyncBuilder = createSynchronizationBuilder({
-            opensearch: this.openSearchClient.use(),
+            openSearchClient: this.openSearchClient.use(),
             timer: this.controller.runtime
         });
         /* We need to find the items we have in the Elasticsearch but not in the DynamoDB-Elasticsearch table. */
